Split MCP integration tests from workspace model
